requirements.txt
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
crispy-bootstrap5==0.7 ; python_version >= "3.9" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.9" and python_version < "4.0"
django-filter==23.2 ; python_version >= "3.9" and python_version < "4.0"
django-htmx==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
django-tables2==2.6.0 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.3 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
